merge mvp_auth

+++ src/user/user.service.ts
@@ -1,44 +1,75 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+
+import { ConflictException, Injectable, NotFoundException, } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { Prisma, User } from '@prisma/client';
 import { IsEmail } from 'class-validator';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class UserService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(dto: CreateUserDto) {
-    const user = await this.prisma.user.findUnique({
-      where: {
-        email: dto.email,
-      },
+  async create(createUserDto: CreateUserDto) {
+    const existingUser = await this.prisma.user.findUnique({
+      where: { email: createUserDto.email },
     });
+    if (existingUser) {
+      throw new ConflictException('Usuário já cadastrado com esse email');
+    }
+    const hashedPassword = await bcrypt.hash(createUserDto.senha, 10);
 
-    if (user) throw new ConflictException('email duplicado');
-
-    const newUser = await this.prisma.user.create({
+    const user = await this.prisma.user.create({
       data: {
-        nome: dto.nome,
-        email: dto.email,
-        senha: dto.senha,
-        curso: dto.curso,
-        departamento: dto.departamento,
-        foto_perfil: dto.foto_perfil,
+        nome: createUserDto.nome,
+        email: createUserDto.email,
+        senha: hashedPassword,
+        curso: createUserDto.curso,
+        departamento: createUserDto.departamento,
+        foto_perfil: createUserDto.foto_perfil,
       },
     });
     return newUser;
   }
 
   async findAll() {
-    return await this.prisma.user.findMany();
+    return await this.prisma.user.findMany({
+      select: {
+        id: true,
+        nome: true,
+        email: true,
+        curso: true,
+        departamento: true,
+        Avaliacoes: true,
+        Comentarios: true,
+        createdAt: true,
+        updatedAt: true,
+      },
+    });
   }
 
   async findOne(id: number) {
+    const isValidId = await this.prisma.user.findUnique({ where: { id } });
+    if (!isValidId) {
+      throw new NotFoundException(
+        `O usuário com o id ${id} não foi encontrado`,
+      );
+    }
     return await this.prisma.user.findUnique({
-      where: { 
-        id : id, 
+      where: {
+        id: id,
+      },
+      select: {
+        id: true,
+        nome: true,
+        email: true,
+        curso: true,
+        departamento: true,
+        Avaliacoes: true,
+        Comentarios: true,
+        createdAt: true,
+        updatedAt: true,
       },
       include: {
         Avaliacoes: {
@@ -66,25 +97,63 @@ export class UserService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const isValidId = await this.prisma.user.findUnique({ where: { id } });
+    if (!isValidId) {
+      throw new NotFoundException(
+        `O usuário com o id ${id} não foi encontrado`,
+      );
+    }
+
+    const hashedPassword = await bcrypt.hash(updateUserDto.senha, 10);
+
     return await this.prisma.user.update({
-      where: { 
-        id : id, 
+      where: {
+        id: id,
       },
       data: {
         nome: updateUserDto.nome,
         email: updateUserDto.email,
-        senha: updateUserDto.senha,
+        senha: hashedPassword,
         curso: updateUserDto.curso,
         departamento: updateUserDto.departamento,
         foto_perfil: updateUserDto.foto_perfil,
+      },
+      select: {
+        id: true,
+        nome: true,
+        email: true,
+        curso: true,
+        departamento: true,
+        Avaliacoes: true,
+        Comentarios: true,
+        createdAt: true,
+        updatedAt: true,
       },
     });
   }
 
   async remove(id: number) {
+    const isValidId = await this.prisma.user.findUnique({ where: { id } });
+    if (!isValidId) {
+      throw new NotFoundException(
+        `O usuário com o id ${id} não foi encontrado`,
+      );
+    }
+
     return await this.prisma.user.delete({
-      where: { 
-        id : id, 
+      where: {
+        id: id,
+      },
+      select: {
+        id: true,
+        nome: true,
+        email: true,
+        curso: true,
+        departamento: true,
+        Avaliacoes: true,
+        Comentarios: true,
+        createdAt: true,
+        updatedAt: true,
       },
     });
   }
